Correccion RegisterAcivity

+++ app/src/main/java/com/alex/deu/RegisterActivity.java
@@ -1,7 +1,6 @@
 package com.alex.deu;
 
 import android.content.Context;
-import android.os.Environment;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.hardware.Sensor;
@@ -12,12 +11,11 @@ import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
-
 import java.io.File;
 import java.io.FileOutputStream;
 import java.util.Calendar;
 
-public class RegisterActivity extends AppCompatActivity implements SensorEventListener{
+public class RegisterActivity extends AppCompatActivity implements SensorEventListener {
 
     private static final String TAG = "RegisterActivity";
 
@@ -25,10 +23,8 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
     private Sensor mAcce, mGyro;
     private TextView textData;
 
-    private static final int m = 600, n = 3;
-    // con 600 muestras cada 200ms tenemos 2 minutos de datos
-
-    //private float[][] acc_data = new float[m][n];
+    private static final int m = 1200, n = 3;
+    // con 600 muestras (m) cada 200ms, tenemos 2 minutos de datos
 
     private float[][] acc_data = null;
     private int acc_data_line = 0;
@@ -36,9 +32,9 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
     private float[][] gyr_data = null;
     private int gyr_data_line = 0;
 
-
-
     public void startRegister(View view) {
+        Toast toast = Toast.makeText(this, "START: Registrando actividad de sosnsores", Toast.LENGTH_SHORT);
+        toast.show();
 
         //SENSOR_DELAY_NORMAL 200.000 microseconds
         //SENSOR_DELAY_UI 60.000 microseconds
@@ -49,20 +45,21 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
         acc_data = new float[m][n];
         gyr_data = new float[m][n];
 
-        if (mAcce!= null) {
+        if (mAcce != null) {
             acc_data_line = 0;
             sensorManager.registerListener(this, mAcce, SensorManager.SENSOR_DELAY_NORMAL);
             Log.d(TAG, "startRegister: Registrando el sensor acelerometro");
-        }else{
+        } else {
             Log.d(TAG, "startRegister: Accelerometro NO DISPONIBLE");
         }
-        if (mGyro!= null) {
+        if (mGyro != null) {
             gyr_data_line = 0;
             sensorManager.registerListener(this, mGyro, SensorManager.SENSOR_DELAY_NORMAL);
             Log.d(TAG, "startRegister: Registrando el sensor giroscopio");
-        }else{
+        } else {
             Log.d(TAG, "startRegister: Giroscopio NO DISPONIBLE");
         }
+
     }
 
     public void stopRegister(View view) {
@@ -72,29 +69,29 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
         //Log.d(TAG, "stopRegister: "+ dataToString(acc_data));
         //textData.setText("Datos guardados");
 
-        if(acc_data != null) {
-            makeDataFile(dataToString(acc_data),1);
+        if (acc_data != null) {
+            makeDataFile(dataToString(acc_data), 1);
             acc_data = null;
         }
-        if (gyr_data != null){
-            makeDataFile(dataToString(gyr_data),2);
+        if (gyr_data != null) {
+            makeDataFile(dataToString(gyr_data), 2);
             gyr_data = null;
         }
 
-
         //Mostrar lista de ficheros guardados
         String text = "";
-        int i = 0;
-        for(i=0;i<fileList().length;i++){
-            text = text + fileList()[i] + "\n";
+        int i;
+        for (i = 0; i < fileList().length; i++) {
+            text += fileList()[i] + "\n";
         }
         textData.setText(text);
 
     }
 
-    public void makeDataFile(String data, int tipo){
+    public void makeDataFile(String data, int tipo) {
         // Tipo 1 == acelerometro
         // Tipo 2 == giroscopio
+        // Tipo 3 == acelerometro lineal
 
 
         Log.d(TAG, "stopRegister: Creating FILE");
@@ -110,15 +107,15 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
         int day = c.get(Calendar.DAY_OF_MONTH);
         int month = c.get(Calendar.MONTH);
         int year = c.get(Calendar.YEAR);
-        String date = year + "-" + month + "-" + day +"_"+ hour + "-" + min + "-" + sec;
+        String date = year + "-" + month + "-" + day + "_" + hour + "-" + min + "-" + sec;
         String filename = "";
 
         if (tipo == 1) {
             filename = "acc_" + date + ".txt";
-
-        }else if (tipo==2){
+        } else if (tipo == 2) {
             filename = "gyr_" + date + ".txt";
         }
+
         //Crear fichero
         File directory = this.getFilesDir();
         File file = new File(directory, filename);
@@ -133,34 +130,52 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
             e.printStackTrace();
         }
 
-        Log.d(TAG, "stopRegister: Creating FILE... DONE"+directory+"/"+filename);
+        Log.d(TAG, "stopRegister: Creating FILE... DONE" + directory + "/" + filename);
     }
 
-    public String dataToString(float datos[][]){
+    public String dataToString(float datos[][]) {
         String text = "";
-        int i, j;
-        for (i=0;i < m;i++) {
-            for (j = 0; j < n; j++) {
-                text += Float.toString(datos[i][j]) + "\t";
-                if (j == 2) {
-                    text += "\n";
+        int i;
+        for (i = 0; i < m; i++) {
+
+            if (datos[i][0] == 0f) {
+                return text;
+            } else {
+                if (i == 0) {
+                    text += Float.toString(datos[i][0]) + "\t" +
+                            Float.toString(datos[i][1]) + "\t" +
+                            Float.toString(datos[i][2]);
+                } else {
+                    text += "\n" + Float.toString(datos[i][0]) + "\t" +
+                            Float.toString(datos[i][1]) + "\t" +
+                            Float.toString(datos[i][2]);
                 }
             }
         }
         return text;
     }
 
-
     public void deleteLastFile(View view) {
-        Toast toast = Toast.makeText(this, "Borando...", Toast.LENGTH_SHORT);
+        Toast toast;
+        //toast = Toast.makeText(this, "Borando...", Toast.LENGTH_SHORT);
         //toast.show();
-
+        if(fileList().length > 0) {
             int last = fileList().length - 1;
             this.deleteFile(fileList()[last]);
 
-
-        toast = Toast.makeText(this, "Archivo eliminado", Toast.LENGTH_SHORT);
-        toast.show();
+            toast = Toast.makeText(this, "File deleted", Toast.LENGTH_SHORT);
+            toast.show();
+        }else{
+            toast = Toast.makeText(this, "fileList() Empty", Toast.LENGTH_SHORT);
+            toast.show();
+        }
+        //Mostrar lista de ficheros guardados
+        String text = "";
+        int i;
+        for (i = 0; i < fileList().length; i++) {
+            text += fileList()[i] + "\n";
+        }
+        textData.setText(text);
     }
 
     @Override
@@ -178,6 +193,7 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
 
         textData = findViewById(R.id.textView_accData);
     }
+
     @Override
     public void onAccuracyChanged(Sensor sensor, int accuracy) {
 
@@ -188,7 +204,7 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
         Sensor sensor = event.sensor;
         if (sensor.getType() == Sensor.TYPE_ACCELEROMETER) {
             Log.d(TAG, "Acce Changed: X:" + event.values[0] + " Y:" + event.values[1] + " Z:" + event.values[2]);
-            if(acc_data_line < m) {
+            if (acc_data_line < m) {
                 acc_data[acc_data_line][0] = event.values[0];
                 acc_data[acc_data_line][1] = event.values[1];
                 acc_data[acc_data_line][2] = event.values[2];
@@ -197,7 +213,7 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
         }
         if (sensor.getType() == Sensor.TYPE_GYROSCOPE) {
             Log.d(TAG, "Gyro Changed: X:" + event.values[0] + " Y:" + event.values[1] + " Z:" + event.values[2]);
-            if(gyr_data_line < m) {
+            if (gyr_data_line < m) {
                 gyr_data[gyr_data_line][0] = event.values[0];
                 gyr_data[gyr_data_line][1] = event.values[1];
                 gyr_data[gyr_data_line][2] = event.values[2];
@@ -207,7 +223,10 @@ public class RegisterActivity extends AppCompatActivity implements SensorEventLi
     }
 
 
-
-
-
+    @Override
+    protected void onPause() {
+        super.onPause();
+        sensorManager.unregisterListener(this, mAcce);
+        sensorManager.unregisterListener(this, mGyro);
+    }
 }
